Refactor SSE4.2 detection for hardware CRC32-C and software fallback
