+++ src/crc32c/crc32c.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <re.h>
 
-uint32_t rawrtc_crc32c(uint32_t crc, void const* buffer, size_t length);
-
 void rawrtc_crc32c_init(void);
+
+uint32_t rawrtc_crc32c(void const* buffer, size_t length);
