+++ src/crc32c/meson.build
@@ -1,1 +1,4 @@
 sources += files('crc32c.c')
+if have_cpuid
+    sources += files('sse42.c')
+endif
